+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 ## Changelog
 
 #### Unreleased
-- [story] [Add BusinessAppsClient](https://www.pivotaltracker.com/story/show/166155315)
+
+#### 1.2.0
+- [story] [Add BusinessAppClient](https://www.pivotaltracker.com/story/show/166155315)
 - [fix] [Passed params in nested arrays are not being squashed](https://www.pivotaltracker.com/story/show/166155374)
 - [new] added `DoubleOptInExecutionClient` and `DoubleOptInRequestClient` clients